+++ CHANGELOG.md
@@ -6,6 +6,13 @@
 All notable changes to this project will be documented here.
 
 
+[Added] ```getMutualGroup``` method
+
+[Added] ```onGetMutualGroups``` callback
+
+[Added] ```getContacts``` => add username to RequestGetContact
+
+
 ## Version [0.9.9.0]
 [BugFixed] Failure to receive history in specific situations